requirements.txt
asgiref==3.5.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cloudinary==1.33.0
cryptography==39.0.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==2.0.0
Django==4.1.5
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-oauth-toolkit==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-social-auth==0.1.0
drf-social-oauth2==1.2.1
drf-spectacular==0.26.1
ecdsa==0.18.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
jwcrypto==1.4.2
oauthlib==3.2.2
Pillow==9.2.0
psycopg2-binary==2.9.6
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
typing_extensions==4.6.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.4.0
wrapt==1.14.1
